requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contractions==0.1.73
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
fastapi-slim==0.111.0
filelock==3.15.4
fsspec==2024.6.0
h11==0.14.0
idna==3.7
inexactsearch==1.0.2
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==5.1.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pluggy==1.5.0
preshed==3.0.9
pyahocorasick==2.1.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyspellchecker==0.8.1
pytest==8.2.2
pytest-mock==3.14.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
setuptools==70.1.0
shellingham==1.5.4
silpa_common==0.3
smart-open==7.0.4
sniffio==1.3.1
soundex==1.1.3
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
sympy==1.12.1
textsearch==0.0.24
thinc==8.2.5
torch==2.3.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
